+++ ghost/core/package.json
@@ -92,7 +92,6 @@
     "@tryghost/html-to-plaintext": "0.0.0",
     "@tryghost/http-cache-utils": "0.1.17",
     "@tryghost/i18n": "0.0.0",
-    "@tryghost/identity-token-service": "0.0.0",
     "@tryghost/image-transform": "1.3.0",
     "@tryghost/importer-handler-content-files": "0.0.0",
     "@tryghost/in-memory-repository": "0.0.0",
