+++ ghost/core/test/unit/server/services/identity-tokens/IdentityTokenService.test.ts
@@ -1,5 +1,5 @@
 import assert from 'assert/strict';
-import {IdentityTokenService} from '../src';
+import {IdentityTokenService} from '../../../../../core/server/services/identity-tokens/IdentityTokenService';
 import {JWK} from 'node-jose';
 import {verify} from 'jsonwebtoken';
