--- ghost/identity-token-service/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@tryghost/identity-token-service",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/identity-token-service",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
-  "scripts": {
-    "dev": "tsc --watch --preserveWatchOutput --sourceMap",
-    "build": "tsc",
-    "build:ts": "tsc",
-    "prepare": "tsc",
-    "test:unit": "NODE_ENV=testing c8 --src src --all --check-coverage --100 --reporter text --reporter cobertura mocha -r ts-node/register './test/**/*.test.ts'",
-    "test": "yarn test:types && yarn test:unit",
-    "test:types": "tsc --noEmit",
-    "lint:code": "eslint src/ --ext .ts --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .ts --cache"
-  },
-  "files": [
-    "build"
-  ],
-  "devDependencies": {
-    "c8": "10.1.2",
-    "mocha": "10.8.2",
-    "node-jose": "2.2.0",
-    "sinon": "19.0.2",
-    "ts-node": "10.9.2",
-    "typescript": "5.6.3"
-  },
-  "dependencies": {
-    "jsonwebtoken": "8.5.1"
-  }
-}
