--- ghost/identity-token-service/src/index.ts
@@ -1,1 +1,0 @@
-export * from './IdentityTokenService';
